fix(release): poll incomplete registry metadata
